Updated PatrolEnemySpawner to correct math
Basically all of my math was wrong, so this is a fixed version. Debug statements can be uncommented for troubleshooting.
Added a delay in calling StartPatrolling() on the Enemy in case the Enemy is not quite ready to receive input like that.
Updated Stag to be faster when patrolling.
Updated Spawner's .h file so only the proper variables are exposed and all exposed variables have comments