Added Beginning of Merchant & Item purchases
Added the concept of a merchant that can sell and purchase items
Overview: when a user Interacts (hits 'e') the code first looks for an overlapping item. If none is found, it does a line trace from the Camera (center of the screen). If a class that implements the Interactable interface, Interact() is called on it. In the case of a Merchant the Merchant's version of Interact() opens the Merchant display where the user may buy or sell items. The Merchant display widget closes if the user hits 'e' again or 'esc'
When interacting, added line tracing to pawns in order to interact with Merchants
Merchant: changed NPCs from AActors to ACharacters to support animating them; moved the promise of Interaction from NPCs (not all NPCs are interactable) to Merchants specifically
Items: Added the concept that items can cost amounts that aren't just gold but actual different items; the way I've implemented this will change soon