Added an inventory
Added an inventory and updated the Item Datatable to be based on a c++ struct.
The SC owns the Inventory and adds items to it.
The inventory can add either (a) exactly as much as was requested and no less, or (b) as much as possible out of what was requested. If the inventory can't hold all that was requested, the user will throw the remaining items on the ground.
Also updated the SlashCharacter's BP to not be able to kick the items around because it was annoying.