Added viewing inventory in HUD
Updated the Tab key functionality to also show the inventory.
The numbers sum out: the length of the inventory widget as shown in the overlay = (slot size + padding * 2 + padding of inventory widget) * # slots (aka 10). Changing the width of the inventory widget or any of these numbers should take this into consideration
Overall for the design I followed Ryan Laley's tutorials, though of course the way I store / retrieve / update the information in the inventory is different. Also, none of my code is replicated.
The user cannot currently interact with the inventory - that comes later. This commit is just about seeing the inventory.
Known issue: if the user acquires an item while they have the inventory open, the view will not be updated.