Added moving items in the Inventory UI
Items can be transferred between slots.
If the two items have different ItemIDs the items are Swapped.
If the two items have the same ItemIDs the items are Stacked.
This works in only the local inventory at the moment. There are currently no non-local inventory sources in the game.
Might have to create a different Inventory type altogether for non-local inventory sources as there are a lot of bits of functionality that simply don't matter when the Inventory isn't owned by the Player Character.