Updated to handle moving items across Inventories
Created Container class which is an Actor which is Interactable. It has an Inventory which the user can hit 'e' to bring up and then drag and drop items into the container.
Future enhancements include an "Take All" button on the container to bring all items into the user's inventory. Also, larger containers.
Also, I tried to write login in Blueprints but it was awful and I hated it so I created a backing InventoryWithSlot widget and that was MUCH better.