Added ctrl+click & showing picked up items
Now when an item is picked up, the item name + amount of the item picked up is displayed on the screen. When the timer expires it fades out. If another item is picked up before the fade out is finished, the fadeout is cancelled and the new item name is displayed.
Items can be dropped from inventories using ctrl+click. The item is spawned and tossed onto the ground in front of the player.
Items can be moved from the player's inventory to an open container using ctrl+click. If there are more items in the stack that the user is ctrl+clicking on than can fit in the container, the remainder of the stack will stay in the user's inventory, in the same slot.
Items can be moved from containers into the player's inventory using ctrl+click.