Added ability to Use items
Added ability for users to right-click on items in their inventory and 'use' them. Items now have use types and I've implemented two of them: "direct" use, where the item has the information in their data table about how the item is "applied". Example: potions. A potion simply increments Health a certain Amount (and since Health is considered an Item, I simply call "AddItem" on it and that works great). I imagine food would work similarly. Also implemented is Weapons: they are equipped, and there are several pieces of logic there.
For weapons I handle the case where a weapon is equipped while no weapon is currently equipped;
A weapon is equipped while there is a currently equipped weapon;
A weapon is equipped from the user's inventory;
A weapon is equipped from a container;
A weapon is equipped from a container to replace an existing weapon and the user's inventory is full (this causes the weapon to be tossed).
Note: when a weapon is equipped from the user's inventory, the currently equipped weapon goes into the same slot the new weapon was equipped from. But when a weapon is equipped from a container, it is added to the first available slot (AddLenient()) in the user's inventory.
When an item is used, the item may want a sound to be played, particles to be played, or an animation to be played. Eventually an animation will be required, but right now in the interest of time I don't even consider animations; I simply set a timer. Then the "finishing" function which actually applies the change is called. Eventually it will be an anim notify.
Removed the Weapon Equip code: now when a weapon is interacted with it will go into the user's inventory and they can equip it from there, like any other item.
Added an ActionState for the character and a walkspeed variable for that state. Thank the gods for my previously refactored ActionState updater function! So useful.
Moved Item spawning & tossing in the SlashCharacter into functions.