Implemented Merchant buy & sell
Allowed buying and selling from the Merchant. Added "Buy" and "Sell" buttons, displayed the clicked item as the Selected item above the lists. Added SlashCharacter functionality to accept the items that will be added and those that will be removed and allowed the user's Gold to be updated when items were purchased or sold.
Added a handful more items to the Merchant shop to be sold.
Turns out I don't need Inventory->AddStrict() after all. Removed it.