First half of Quests feature
Created a Quest Component and added it to the SlashCharacter.
Added an Enum with all of the ObjectiveTypes I'll need for now.
Added an EnemyID and created an Enemy DataTable for ID association.
Created the default quest and its objectives. Auto-bestowed the quest to the PC.
Created a subsystem that accepts events to its CreateEvent function. Called that function in several places, and verified they're coming into the QuestComponent as they should.
Also initialized a bunch of values in the AttributesComponent. It's confusing when the BP zeros them out when I expect them to be 100.