Multiple bug fixes
Updated enemy health to be more realistic vs the damage output of the Character's weapons.
Created a GameMode for the Del character.
Found I accidentally cast the PC to a BP_SlashCharacter instead of a SlashCharacter in the TabWidget, causing Del to not be able to request a levelup.
Updated the IMC with valid bindings to gamepad keys for the various mappings I have.
Fixed an issue with not being able to purchase the DoomBringer if the user had exactly 200 Gold (the exact price of the item).
The PC was able to hit the wolf model multiple times after the wolf died, causing multiple souls to spawn. I modified the die function to update the collision response channel of the enemy's mesh to prevent that.