Many tweaks and added the Roxanne character
Added a Main menu and some sub-menus
Updated Merchant Widget Menu: plays sounds when items are bought or sold, shows only the items that can be sold and are in the user's inventory and the amounts of those items.
Added new Roxanne character - BP, ABPs, AMs
Changed UseItem() from just waiting 2 seconds to playing an animation (if one exists), then the animation calls FinishUseItem()
Added an effect that is played on Leveling Up
Added day/night cycles and some music that plays when the day changes to night and vice versa
Added a message when the user dies and the user is now taken to the main menu after a few seconds.
Added a mining pick that appears when the player mines ore. If the player is carrying a weapon, the weapon is hidden until the mining animation is done.
Modified the AttributeComponent->GetPercentToNextLevel() to return a value between 0 and 1 (it might have been causing the game to crash in some circumstances)