Update Aggro/Combat state based on Sphere instead of PawnSensing
I wanted more control over enemies aggroing on the PC and other enemies, so while I am keeping the PawnSensing component because I want to use it later for hearing sound, I am now using a USphereComponent to keep track of the main mob close to the Enemy, and while there is an AggroTarget (an enemy within the AggroSphere radius) the enemy tests in Tick whether it can see the AggroTarget based on user-controlled Left and Right Angle variables.
The test for whether a player is within the Angles is taken from the BaseCharacter's DirectionalHitReact function.
Got rid of CombatRadius variable in favor of using AggroRadius, which largely means the same thing. AggroRadius is not set by the player, but determined by the radius of the AggroSphere, which is set by the player.
Updated the CombatTarget variable to match the AggroTarget variable - they can only hold ABaseCharacters instead of AActors.
Also now the enemy interps toward the player, so the player should never be able to sit behind an enemy and kill them in safety.