Added a configurable threshold for a mob hit reacting
Updated the threshold at which the ABaseCharacter should hit react. Previously any hit was causing any character to hit react.
Now, default is when the damage dealt is >= 25% of the character's current health, but is configurable in blueprints. A character hit for 2 should hit react if their current health is <= 8, and not hit react if their current health is > 8. Overall this will prevent the player from stunlocking enemies, and prevent enemies from stunlocking the player.