Improved Detecting Actors for pickup
Updated collision on enemies and swords.
Updated LineTrace for picking up interactable objects to trace by the Visibility Channel rather than for entire objects (it was causing the line trace to detect a hit on the enemy AggroSpheres which was not at all what I wanted).
Updated Interaction code to limit picking up items to when the player is Unoccupied or Sprinting.